dev_test/requirements.txt
streamlit==0.89.0
pandas==1.3.3
psycopg2-binary==2.9.1
plotly==5.3.1
requests==2.26.0
